# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Tereka personal finance intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type Currency = typeof Currency[keyof typeof Currency];


export const Currency = {
  UGX: 'UGX',
  KES: 'KES',
  TZS: 'TZS',
  RWF: 'RWF',
  USD: 'USD',
} as const;

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  income: 'income',
  expense: 'expense',
} as const;

export type AccountType = typeof AccountType[keyof typeof AccountType];


export const AccountType = {
  cash: 'cash',
  mobile_money: 'mobile_money',
  bank: 'bank',
  savings: 'savings',
  other: 'other',
} as const;

export interface Category {
  id: string;
  name: string;
  type: TransactionType;
  icon: string;
  isDefault: boolean;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  type: TransactionType;
  icon: string;
}

export interface Account {
  id: string;
  name: string;
  type: AccountType;
  currency: Currency;
  openingBalance: number;
  balance: number;
  isActive: boolean;
}

export interface AccountInput {
  /** @minLength 1 */
  name: string;
  type: AccountType;
  currency: Currency;
  /** @minimum 0 */
  openingBalance: number;
}

export interface AccountUpdate {
  /** @minLength 1 */
  name?: string;
  type?: AccountType;
  currency?: Currency;
  /** @minimum 0 */
  openingBalance?: number;
  isActive?: boolean;
}

export interface Transaction {
  id: string;
  type: TransactionType;
  amount: number;
  currency: Currency;
  accountId: string;
  accountName: string;
  categoryId: string;
  categoryName: string;
  description: string;
  /** @nullable */
  notes: string | null;
  transactionDate: string;
}

export interface TransactionInput {
  type: TransactionType;
  /** @exclusiveMinimum 0 */
  amount: number;
  currency: Currency;
  accountId: string;
  categoryId: string;
  /** @minLength 1 */
  description: string;
  notes?: string;
  transactionDate: string;
}

export interface TransactionUpdate {
  type?: TransactionType;
  /** @exclusiveMinimum 0 */
  amount?: number;
  currency?: Currency;
  accountId?: string;
  categoryId?: string;
  /** @minLength 1 */
  description?: string;
  notes?: string;
  transactionDate?: string;
}

export type BudgetPeriod = typeof BudgetPeriod[keyof typeof BudgetPeriod];


export const BudgetPeriod = {
  monthly: 'monthly',
} as const;

export type BudgetStatus = typeof BudgetStatus[keyof typeof BudgetStatus];


export const BudgetStatus = {
  on_track: 'on_track',
  warning: 'warning',
  exceeded: 'exceeded',
} as const;

export interface Budget {
  id: string;
  categoryId: string;
  categoryName: string;
  amount: number;
  spent: number;
  currency: Currency;
  period: BudgetPeriod;
  percentageUsed: number;
  status: BudgetStatus;
}

export type BudgetInputPeriod = typeof BudgetInputPeriod[keyof typeof BudgetInputPeriod];


export const BudgetInputPeriod = {
  monthly: 'monthly',
} as const;

export interface BudgetInput {
  categoryId: string;
  /** @exclusiveMinimum 0 */
  amount: number;
  currency: Currency;
  period: BudgetInputPeriod;
}

export type BudgetUpdatePeriod = typeof BudgetUpdatePeriod[keyof typeof BudgetUpdatePeriod];


export const BudgetUpdatePeriod = {
  monthly: 'monthly',
} as const;

export interface BudgetUpdate {
  categoryId?: string;
  /** @exclusiveMinimum 0 */
  amount?: number;
  currency?: Currency;
  period?: BudgetUpdatePeriod;
}

export type FinancialGoalStatus = typeof FinancialGoalStatus[keyof typeof FinancialGoalStatus];


export const FinancialGoalStatus = {
  active: 'active',
  completed: 'completed',
  paused: 'paused',
} as const;

export interface FinancialGoal {
  id: string;
  name: string;
  targetAmount: number;
  currentAmount: number;
  currency: Currency;
  targetDate: string;
  status: FinancialGoalStatus;
  percentageComplete: number;
  remainingAmount: number;
}

export type GoalInputStatus = typeof GoalInputStatus[keyof typeof GoalInputStatus];


export const GoalInputStatus = {
  active: 'active',
  completed: 'completed',
  paused: 'paused',
} as const;

export interface GoalInput {
  /** @minLength 1 */
  name: string;
  /** @exclusiveMinimum 0 */
  targetAmount: number;
  /** @minimum 0 */
  currentAmount: number;
  currency: Currency;
  targetDate: string;
  status: GoalInputStatus;
}

export type GoalUpdateStatus = typeof GoalUpdateStatus[keyof typeof GoalUpdateStatus];


export const GoalUpdateStatus = {
  active: 'active',
  completed: 'completed',
  paused: 'paused',
} as const;

export interface GoalUpdate {
  /** @minLength 1 */
  name?: string;
  /** @exclusiveMinimum 0 */
  targetAmount?: number;
  /** @minimum 0 */
  currentAmount?: number;
  currency?: Currency;
  targetDate?: string;
  status?: GoalUpdateStatus;
}

export interface CategorySpend {
  categoryName: string;
  amount: number;
  percentage: number;
  color: string;
}

export interface MonthlyComparison {
  month: string;
  income: number;
  expenses: number;
}

export interface DashboardSummary {
  currency: Currency;
  totalBalance: number;
  monthlyIncome: number;
  monthlyExpenses: number;
  remainingBudget: number;
  balanceChange: number;
  spendingByCategory: CategorySpend[];
  incomeVsExpenses: MonthlyComparison[];
  recentTransactions: Transaction[];
  budgetStatus: Budget[];
  goalProgress: FinancialGoal[];
}

export type FinancialInsightTone = typeof FinancialInsightTone[keyof typeof FinancialInsightTone];


export const FinancialInsightTone = {
  positive: 'positive',
  neutral: 'neutral',
  attention: 'attention',
} as const;

export interface FinancialInsight {
  id: string;
  title: string;
  body: string;
  tone: FinancialInsightTone;
  createdAt: string;
}

export interface Conversation {
  id: string;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface ConversationInput {
  title?: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: string;
  conversationId: string;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export type ProfileTheme = typeof ProfileTheme[keyof typeof ProfileTheme];


export const ProfileTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface Profile {
  id: string;
  fullName: string;
  email: string;
  country: string;
  preferredCurrency: Currency;
  theme: ProfileTheme;
}

export type ProfileUpdateTheme = typeof ProfileUpdateTheme[keyof typeof ProfileUpdateTheme];


export const ProfileUpdateTheme = {
  light: 'light',
  dark: 'dark',
  system: 'system',
} as const;

export interface ProfileUpdate {
  /** @minLength 1 */
  fullName?: string;
  country?: string;
  preferredCurrency?: Currency;
  theme?: ProfileUpdateTheme;
}

export type SearchParameter = string;

export type TransactionTypeParameter = TransactionType;

export type CategoryIdParameter = string;

export type GetTransactionsParams = {
search?: SearchParameter;
type?: TransactionTypeParameter;
categoryId?: CategoryIdParameter;
};

export type GetCategoriesParams = {
type?: TransactionTypeParameter;
};